+++ src/lib/st-search/demo/st-search-demo.ts
@@ -82,7 +82,6 @@ export class StSearchDemoComponent {
    }
 
    searchWithAutocompleteSearch(searchValue: string): void {
-      console.log('search');
       this.searchedAutocomplete = searchValue;
       this.filter(searchValue);
    }